Committing a series of scripts that will generate 23B rows of synthetic, production-ish data in elastic search.
The script pulls from a CSV that represents all of the datasets with existing text columns,
and the count of the text columns within that dataset.  For each text column it then
generates a new fake string per row in the dataset. The length of that fake string is
a random string between 1 and 255.  That string is then batched up with a bunch of
other updates and then bulk inserted whenever the batches exceed 15 mb.