EN-17472 Ensure suggestions are unique
This fundamentally changes the way we serve suggestions. Rather than using the in-memory completion
suggester which no longer returns unique suggestions, we use an autocomplete subfield consisting of
edge ngrams that are created at index-time. This has the advantage of no longer pinning the
suggestion finite state transducer data structure to memory, with a corresponding trade-off in
performance. (Details on the impact on performance forthcoming.)

The simplifies the codebase and normalizes it with an eye towards consistency with Cetera. A
smattering of changes for consistency:

- moved ES settings and mappings out of reference configuration file into JSON resource files
- added scripts for setting up index and creating index alias

Lots of cruft was removed as a result of removing the no longer necessary CompletionAnalyzer. All
analysis is defined directly in the ES settings file.

There is a subtle change in scoring. When no text is provided to the endpoint, rather than
returning a meaningless score, we return the frequency of the values. When text is provided, the
score is the vanilla relevance score returned by Elasticsearch.

In order to return unique suggestion results, we execute a terms aggregation along with the
query (dataset, column, and optional text constraints). When query text is provided, we pass along
a max subaggregation that gets the original scores from the returned documents. Using a script to
get the max score breaks unit tests due to the fact that Elasticsearch no longer supports local ES
nodes for unit testing. Future work: break out functional/integration tests into a separate
submodule that uses the official ES integration testing framework.

Eyeballed results locally on staging corpus. Unit tests pass.