Add new mapping to avoid deduplicating suggestions via aggregation
With the initial upgrade to Elasticsearch 5, we moved away from Elasticsearch's completion
suggester for a couple of reasons:

1. it's not scalable (ie. the memory requirements make it cost prohibitive)
2. it no longer auto dedupes

The first change was to replace the completion suggester via the use of a new autocomplete field in
the mapping and a new edge ngram analyzer. The second change -- specifically to address the
deduplication issue -- was to introduce an additional aggregation query as part of the suggestions
request to ensure that the suggested values are unique.

After some performance testing, despite an expected impact on performance, the individual query
performance looked reasonable (<1 sec in the aggregate). Once we deployed to production however, we
saw a different story -- performance degradation was significant in the context of concurrent
requests.

This next revision builds off of that last step. We remove the need for an aggregation to dedupe
suggestions by restructuring the mapping to store column suggestions as (suggestion, count)
pairs. This new mapping is intuitive. We store only the data we need to satisfy the
request. This is possible because the secondary receives deltas with each row
operation. Specifically, with each insert, deletion, or update we can do the appropriate
book-keeping to ensure counts are correct. Storing counts has the added benefit that they can be
used directly for scoring.

As a side effect of this new mapping (and the need for partial updates to counts), some integration
tests no longer worked with the local ES testing node. These tests were moved to a new
spandex-integration-tests subproject which expect that Elasticsearch is running locally.

- [x] unit tests pass
- [x] integration tests pass
- [x] style check passes